requirements.txt
fasttext-wheel==0.9.2
spacy-langdetect 
networkx>=3.1 
torch==2.1.2 
open_clip_torch==2.23.0
pillow 
openai>=1.14.0
diffusers[torch]==0.25.0 
scikit-learn>=1.3.0 